+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 
+## 0.41.0
+* Add support for StepIn and StepOut: StepIn moves execution one character to the right, StepIn to the left
+* Add support for StepInTargets: every word in the stopped line is considered one stack frame; StepInTargets returns targets for every character of the stack frame with the given frameId.
+
 ## 0.40.0
 * Exercise new dynamic debug config API.